@@ -12,12 +12,12 @@ var (
 	defaultSignals         = map[string]syscall.Signal{"DATA": syscall.SIGUSR1, "STATS": syscall.SIGUSR2}
 )
 
-// HasSigNumSupport checks if Keepalived supports --signum command
+// HasSigNumSupport checks if Keepalived supports --signum command.
 func HasSigNumSupport(version *version.Version) bool {
 	return version == nil || version.GreaterThanOrEqual(sigNumSupportedVersion)
 }
 
-// GetDefaultSignal returns default signals for Keepalived
+// GetDefaultSignal returns default signals for Keepalived.
 func GetDefaultSignal(sigString string) syscall.Signal {
 	sig, ok := defaultSignals[sigString]
 	if !ok {
